refactor : 게시물 통합 테스트 디렉토리 변경
