+++ server/src/test/java/com/momo/post/acceptance/step/PostAcceptanceStep.java
@@ -1,4 +1,4 @@
-package com.momo.post.step;
+package com.momo.post.acceptance.step;
 
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
